+++ src/engraving/libmscore/measure.cpp
@@ -1602,7 +1602,7 @@ EngravingItem* Measure::drop(EditData& data)
     case ElementType::BRACKET:
     {
         Bracket* b = toBracket(e);
-        int level = 0;
+        size_t level = 0;
         staff_idx_t firstStaff = 0;
         for (Staff* s : score()->staves()) {
             for (const BracketItem* bi : s->brackets()) {
